requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-restx==1.2.0
flask-limiter==3.5.0
flask-caching==2.1.0
redis==5.0.1
celery==5.3.4
scikit-learn==1.3.0
nltk==3.8.1
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
prometheus-flask-exporter==0.22.4

# Development dependencies
pytest==7.4.0
black==23.7.0
flake8==6.0.0
coverage==7.3.2